update schnorr sources
